Expose v7_compile() to public
