+++ src/main.c
@@ -10,7 +10,10 @@
 #endif
 
 #ifdef V7_EXE
+#define V7_MAIN
+#endif
 
+#ifdef V7_MAIN
 static void show_usage(char *argv[]) {
   fprintf(stderr, "V7 version %s (c) Cesanta Software, built on %s\n",
           V7_VERSION, __DATE__);
@@ -41,20 +44,6 @@ static char *read_file(const char *path, size_t *size) {
   return data;
 }
 
-static void dump_ast(struct v7 *v7, const char *code, int binary) {
-  struct ast ast;
-
-  ast_init(&ast, 0);
-  if (parse(v7, &ast, code, 1) != V7_OK) {
-    fprintf(stderr, "%s\n", "parse error");
-  } else if (binary) {
-    fwrite(ast.mbuf.buf, ast.mbuf.len, 1, stdout);
-  } else {
-    ast_dump(stdout, &ast, 0);
-  }
-  ast_free(&ast);
-}
-
 static void print_error(struct v7 *v7, const char *f, val_t e) {
   char buf[512];
   char *s = v7_to_json(v7, e, buf, sizeof(buf));
@@ -64,16 +53,25 @@ static void print_error(struct v7 *v7, const char *f, val_t e) {
   }
 }
 
-int main(int argc, char *argv[]) {
+/*
+ * V7 executable main function.
+ * `init_func()` is an optional intialization function, aimed to export any
+ * extra functionality into vanilla v7 engine.
+ */
+int v7_main(int argc, char *argv[], void (*init_func)(struct v7 *)) {
   struct v7 *v7 = v7_create();
   int i, show_ast = 0, binary_ast = 0;
   val_t res = v7_create_undefined();
+
+  if (init_func != NULL) {
+    init_func(v7);
+  }
 
   /* Execute inline code */
   for (i = 1; i < argc && argv[i][0] == '-'; i++) {
     if (strcmp(argv[i], "-e") == 0 && i + 1 < argc) {
       if (show_ast) {
-        dump_ast(v7, argv[i + 1], binary_ast);
+        v7_compile(argv[i + 1], binary_ast, stdout);
       } else if (v7_exec(v7, &res, argv[i + 1]) != V7_OK) {
         print_error(v7, argv[i + 1], res);
         res = v7_create_undefined();
@@ -101,7 +99,7 @@ int main(int argc, char *argv[]) {
       if ((source_code = read_file(argv[i], &size)) == NULL) {
         fprintf(stderr, "Cannot read [%s]\n", argv[i]);
       } else {
-        dump_ast(v7, source_code, binary_ast);
+        v7_compile(source_code, binary_ast, stdout);
         free(source_code);
       }
     } else if (v7_exec_file(v7, &res, argv[i]) != V7_OK) {
@@ -118,5 +116,11 @@ int main(int argc, char *argv[]) {
 
   v7_destroy(v7);
   return EXIT_SUCCESS;
+}
+#endif
+
+#ifdef V7_EXE
+int main(int argc, char *argv[]) {
+  return v7_main(argc, argv, NULL);
 }
 #endif
